#Task3 Splitted structure into src and test folders.
