+++ product-service-ts/handler.ts
@@ -1,5 +1,5 @@
 import 'source-map-support/register';
-import { getProductsById } from './handlers/getProductsById';
-import { getProductsList } from './handlers/getProductsList';
+import { getProductsById } from './src/handlers/getProductsById';
+import { getProductsList } from './src/handlers/getProductsList';
 
 export {getProductsById, getProductsList}
