+++ product-service-ts/src/data/products-mock.json
