+++ product-service-ts/src/handlers/getProductsById.ts
@@ -1,5 +1,5 @@
 import { APIGatewayProxyHandler } from 'aws-lambda';
-import products from './products-mock.json';
+import products from '../data/products-mock.json';
 
 export const getProductsById: APIGatewayProxyHandler = async (event) => {
   console.log("getProductsById FN. Event: ", event);
